Add files via upload
The two videos uploaded here show the line follower performance when controlled by a Raspberry Pi 2 board. The "WP_20160711_18_18_55_ProConverted.mp4" video was the first one. As it can be seen, the tracked vehicle performs poorly (small speed, oscillations) but it does not loose the marker. In the video titled "visual_servoing01.mp4", the camera image dimensions were reduced by half. This allows for a higher speed and reduced tracking oscillations.